# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CipherPay Fintech API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bank/resolve:
    post:
      operationId: resolveAccount
      tags: [bank]
      summary: Resolve/verify a bank account number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveAccountInput"
      responses:
        "200":
          description: Account resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolvedAccount"
        "400":
          description: Invalid account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bank/saved-accounts:
    get:
      operationId: getSavedAccounts
      tags: [bank]
      summary: Get user's saved bank accounts
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Saved accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavedAccount"
    post:
      operationId: saveAccount
      tags: [bank]
      summary: Save a bank account for future use
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveAccountInput"
      responses:
        "201":
          description: Account saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavedAccount"
  /bank/saved-accounts/{id}:
    delete:
      operationId: deleteSavedAccount
      tags: [bank]
      summary: Delete a saved bank account
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── AIRTIME ───────────────────────────────────────────────────────────────
...
  /kyc/status:
    get:
      operationId: getKycStatus
      tags: [kyc]
      summary: Get user KYC verification status
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: KYC status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"
  /kyc/submit:
    post:
      operationId: submitKyc
      tags: [kyc]
      summary: Submit KYC information
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycInput"
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"

  # ── TEMPORARY EMAIL ───────────────────────────────────────────────────────
  /temporary-email/inboxes:
    get:
      operationId: listTemporaryInboxes
      tags: [temporaryEmail]
      summary: List the signed-in user's temporary inboxes
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User's active temporary inboxes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemporaryInboxList"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createTemporaryInbox
      tags: [temporaryEmail]
      summary: Create a temporary inbox for the signed-in user
      parameters:
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Inbox created and stored for this account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemporaryInbox"
        "409":
          description: The user already has two inboxes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Temporary email provider unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /temporary-email/inboxes/{id}:
    delete:
      operationId: deleteTemporaryInbox
      tags: [temporaryEmail]
      summary: Remove an inbox and stop its automatic renewal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - name: x-user-id
          in: header
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inbox removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemporaryInboxRemoval"
        "404":
          description: Inbox not found for this user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── DASHBOARD ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ResolveAccountInput:
      type: object
      required: [accountNumber, bankCode]
      properties:
        accountNumber:
          type: string
        bankCode:
          type: string
    ResolvedAccount:
      type: object
      required: [accountName, accountNumber, bankCode]
      properties:
        accountName:
          type: string
        accountNumber:
          type: string
        bankCode:
          type: string
    SaveAccountInput:
      type: object
      required: [accountNumber, accountName, bankCode, bankName]
      properties:
        accountNumber:
          type: string
        accountName:
          type: string
        bankCode:
          type: string
        bankName:
          type: string
    SavedAccount:
      type: object
      required: [id, userId, accountNumber, accountName, bankCode, bankName, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        accountNumber:
          type: string
        accountName:
          type: string
        bankCode:
          type: string
        bankName:
          type: string
        createdAt:
          type: string

    # ── AIRTIME ─────────────────────────────────────────────────────────────
...
    KycStatus:
      type: object
      required: [userId, level, status, submittedAt]
      properties:
        userId:
          type: integer
        level:
          type: integer
        status:
          type: string
        verificationType:
          type: ["string", "null"]
          enum: [basic, advanced, null]
        documentType:
          type: ["string", "null"]
        bvn:
          type: ["string", "null"]
        nin:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
        verifiedAt:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
    KycInput:
      type: object
      required: [verificationType, documentType]
      properties:
        verificationType:
          type: string
          enum: [basic, advanced]
        documentType:
          type: string
          enum: [bvn, nin, passport, drivers_license, national_id, voters_card]
        documentNumber:
          type: string
        bvn:
          type: ["string", "null"]
        nin:
          type: ["string", "null"]

    # ── DASHBOARD ─────────────────────────────────────────────────────────────
...
    TemporaryInbox:
      type: object
      required: [id, address, inboxId, expiresAt, status, createdAt]
      properties:
        id:
          type: integer
        address:
          type: string
        inboxId:
          type: string
        expiresAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, unavailable]
        createdAt:
          type: string
          format: date-time
    TemporaryInboxList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TemporaryInbox"
    TemporaryInboxRemoval:
      type: object
      required: [success]
      properties:
        success:
          type: boolean